Changed the package structure.
