+++ components/org.wso2.carbon.security/src/main/java/org/wso2/carbon/security/internal/config/StoreConfigBuilder.java
@@ -20,9 +20,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.security.internal.CarbonSecurityDataHolder;
 import org.wso2.carbon.security.jaas.util.CarbonSecurityConstants;
-import org.wso2.carbon.security.usercore.config.AuthorizationStoreConfig;
-import org.wso2.carbon.security.usercore.config.CredentialStoreConfig;
-import org.wso2.carbon.security.usercore.config.IdentityStoreConfig;
+import org.wso2.carbon.security.user.core.config.AuthorizationStoreConfig;
+import org.wso2.carbon.security.user.core.config.CredentialStoreConfig;
+import org.wso2.carbon.security.user.core.config.IdentityStoreConfig;
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.introspector.BeanAccess;
