--- components/org.wso2.carbon.security/src/main/java/org/wso2/carbon/security/usercore/exception/AuthenticationFailure.java
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wso2.carbon.security.usercore.exception;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * This class is an extension of {@code Exception} class and used to indicate authentication failures.
- */
-public class AuthenticationFailure extends Exception {
-
-    private Throwable cause;
-    private List<Throwable> failures = new ArrayList<>();
-
-    public AuthenticationFailure(Throwable e) {
-        this.cause = e;
-    }
-
-    public AuthenticationFailure(String message) {
-        super(message);
-    }
-
-    @Override
-    public Throwable getCause() {
-        return this.cause;
-    }
-
-    public void addFailure(Throwable t) {
-        failures.add(t);
-    }
-
-    public List<Throwable> getFailures() {
-        return failures;
-    }
-}
